requirements.txt
requests
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq